requirements.txt
numpy==1.24.4
scipy==1.10.1
mne==1.6.1
h5py==3.10.0
matplotlib==3.7.5
tqdm==4.67.1
psutil==5.9.0
torch==2.1.0
scikit-learn==1.3.2
umap-learn==0.5.6
pyyaml==6.0.2
lightning==2.2.1
omegaconf==2.3.0
hydra-core==1.3.2